Refactor: seperate elements in BoardPage to components
