--- src/components/Drawers/DrawerContainer/index.tsx
@@ -1,17 +1,0 @@
-import * as S from "./style";
-
-interface Props {
-  isOpened: boolean;
-  isOpenedFromLeft: boolean;
-  children: React.ReactNode;
-}
-
-function DrawerContainer({ isOpened, isOpenedFromLeft, children }: Props) {
-  return (
-    <S.Container isOpened={isOpened} isOpenedFromLeft={isOpenedFromLeft}>
-      {children}
-    </S.Container>
-  );
-}
-
-export default DrawerContainer;
