--- src/components/Drawers/SettingDrawer/index.tsx
@@ -1,17 +1,0 @@
-import React from "react";
-import * as S from "./style";
-
-interface Props {
-  onClose: () => void;
-}
-
-function SettingDrawer({ onClose }: Props) {
-  return (
-    <S.Container>
-      SettingDrawer
-      <button onClick={onClose}>x</button>
-    </S.Container>
-  );
-}
-
-export default SettingDrawer;
