--- src/components/Drawers/SettingDrawer/style.ts
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-export const Container = styled.div`
-  //
-`;
