+++ src/pages/BoardPage/index.tsx
@@ -1,10 +1,14 @@
 import React, { useState } from "react";
-import { Breadcrumb, Layout, theme, Button, Drawer } from "antd";
+import { Breadcrumb, Layout, theme, Button } from "antd";
+import Slider from "@components/Slider";
+import Drawer from "@components/Drawer";
+import Header from "@components/Header";
+import Menu from "@components/Menu";
+import Board from "@components/Board";
 
-const { Header, Content, Sider } = Layout;
+const { Content } = Layout;
 
 const BoardPage: React.FC = () => {
-  const [collapsed, setCollapsed] = useState(false);
   const [open, setOpen] = useState(false);
 
   const showDrawer = () => {
@@ -22,46 +26,15 @@ const BoardPage: React.FC = () => {
     token: { colorBgContainer },
   } = theme.useToken();
 
-  const onClickCollapse = () => {
-    setCollapsed(!collapsed);
-  };
-
   return (
     <Layout style={{ minHeight: "100vh" }}>
-      <Header style={{ position: "sticky", top: 0, zIndex: 1, width: "100%", backgroundColor: "white" }}>
-        <div>header</div>
-      </Header>
+      <Header />
       <Layout className="site-layout">
-        <Sider
-          collapsible
-          collapsed={collapsed}
-          onCollapse={(value) => setCollapsed(value)}
-          style={{ backgroundColor: "white" }}
-        >
-          <button onClick={onClickCollapse}>collapse</button>
-        </Sider>
+        <Slider />
         <Content style={containerStyle}>
-          <Breadcrumb style={{ margin: "16px 0" }}>
-            <div>menu</div>
-            <div style={{ marginTop: 16 }}>
-              <Button type="primary" onClick={showDrawer}>
-                Open
-              </Button>
-            </div>
-          </Breadcrumb>
-          <div style={{ padding: 24, minHeight: 360, background: colorBgContainer }}>cards</div>
-          <div>
-            <Drawer
-              title="Basic Drawer"
-              placement="right"
-              closable={false}
-              onClose={onClose}
-              open={open}
-              getContainer={false}
-            >
-              <p>Some contents...</p>
-            </Drawer>
-          </div>
+          <Menu showDrawer={showDrawer} />
+          <Board boardName={"sdf"} />
+          <Drawer open={open} onClose={onClose} />
         </Content>
       </Layout>
     </Layout>
